Created a form with a drop-down menu for student list and course sessions Able to see the session with the students added. It isn't user friendly. Need to create a for loop like in the route sessions to display all sessions with all students Changed user_sessions table to use bigint instead of smallint

+++ portal/schema.sql
@@ -34,8 +34,8 @@ CREATE TABLE course_sessions (
 
 
 CREATE TABLE user_sessions (
-    student_id smallint REFERENCES users(id),
-    session_id smallint REFERENCES course_sessions(id)
+    student_id bigint REFERENCES users(id),
+    session_id bigint REFERENCES course_sessions(id)
 );
 
 CREATE TABLE assignments (
